functions/requirements.txt
# Azure Functions requirements
azure-functions
azure-functions-durable
azure-servicebus>=7.14.0
python-dotenv
openai
anthropic
requests
beautifulsoup4
psycopg2-binary
sqlalchemy
langchain-openai
langchain-anthropic
redis
pydantic>=2.0.0
httpx>=0.25.0

# ML Libraries
statsmodels>=0.14.0
numpy>=1.24.0
ta>=0.11.0
vaderSentiment>=3.3.2
textblob>=0.17.1
pandas>=2.0.0
prophet>=1.1.5
scikit-learn>=1.3.0
